Unidirectional links are now supported by the topology, but disabled by default
In the current version the unidirectional links, although possible, require additional attention while creating the topology and handling the system. In particular, the nodes having an end of an unidirectional links cannot be turned off while the system is running, as they will not reconnect automatically after being turned on again. Since that might cause significant confusion, they are disabled by default and can be enabled by adding a corresponding preprocessor definition