Removes obsolete function connectTopology() from API
Connect topology function has been required to explicitly connect nodes into a topology. Since a few versions ago the nodes continuosly retry to connect into a topology if it has not been yet found upon the start of the node. As a result, the behavior of connectTopology can be performed automatically while starting the node.