Fixes gather registry cleanup that lead to undelivered messages
Awaiting messages in gather registry are not erased on cleanup anymore. That did not save any space at all, but has lead to undelivered messages if cleanup was called too early. Now, cleanup erases the contents of the delivered messages log, which might only lead to delivering the same message twice if the threshold is set to a low value and cleanup is improperly called.