IMessageSigner now obtains a pointer to the topology rather than a reference
It's better for using message signers just for test purpose, since topology can only be instantiated from valid config file and all test applications would have to supply it, even though the signers itself might not use the topology at all. Now it is possible to simply supply a nullptr in test applications, if the topology is not needed.