Communication between crypto manager and main thread is performed via shared queue
The previous communication, i.e. sending cryptography response, has been performed via zmq channels and in some situations might drop messages due to queue overflow, leading to deadlock in client programs. Right now the zmq channel is only used to transmit a single notification byte, and the rest is stored in a JobQueue inside CryptoManager and is removed in the Node main loop